+++ src/components/ComparisonMetric/ComparisonMetric.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import type {Legend} from '../../types';
+
 import styles from './ComparisonMetric.scss';
 import {UpChevron, DownChevron} from './components';
 import type {ComparisonMetricShape} from './types';
@@ -8,29 +10,36 @@ export function ComparisonMetric({
   metric,
   trend,
   accessibilityLabel,
-}: ComparisonMetricShape) {
+  theme,
+}: ComparisonMetricShape & {theme: Legend}) {
   switch (trend) {
     case 'neutral':
       return (
         <span className={styles.NeutralIcon}>
           <span className={styles.VisuallyHidden}>{accessibilityLabel}</span>
-          <span>-</span>
+          <span style={{color: theme.trendIndicator.neutral}}>-</span>
         </span>
       );
       break;
     case 'positive':
       return (
         <span className={styles.PositiveIcon}>
-          <UpChevron accessibilityLabel={accessibilityLabel} />
-          <span>{metric}</span>
+          <UpChevron
+            accessibilityLabel={accessibilityLabel}
+            fill={theme.trendIndicator.positive}
+          />
+          <span style={{color: theme.trendIndicator.positive}}>{metric}</span>
         </span>
       );
       break;
     case 'negative':
       return (
         <span className={styles.NegativeIcon}>
-          <DownChevron accessibilityLabel={accessibilityLabel} />
-          <span>{metric}</span>
+          <DownChevron
+            accessibilityLabel={accessibilityLabel}
+            fill={theme.trendIndicator.negative}
+          />
+          <span style={{color: theme.trendIndicator.negative}}>{metric}</span>
         </span>
       );
   }