+++ src/components/ComparisonMetric/components/UpChevron/UpChevron.tsx
@@ -1,16 +1,15 @@
 import React from 'react';
 
-import {getColorValue} from '../../../../utilities';
 import {XMLNS} from '../../../../constants';
 
 import styles from './UpChevron.scss';
 
 interface Props {
   accessibilityLabel: string;
+  fill: string;
 }
 
-export function UpChevron({accessibilityLabel}: Props) {
-  const fill = getColorValue('positive');
+export function UpChevron({accessibilityLabel, fill}: Props) {
   return (
     <React.Fragment>
       <span className={styles.VisuallyHidden}>{accessibilityLabel}</span>