+++ src/components/LinePreview/LinePreview.tsx
@@ -1,7 +1,7 @@
 import React, {useRef} from 'react';
-import type {SeriesColor} from 'types';
+import type {Color} from 'types';
 
-import {getColorValue, isGradientType, uniqueId} from '../../utilities';
+import {isGradientType, uniqueId} from '../../utilities';
 import type {LineStyle} from '../../types';
 import {LinearGradient} from '../LinearGradient';
 import {XMLNS} from '../../constants';
@@ -14,7 +14,7 @@ import {
 } from './constants';
 
 interface Props {
-  color: SeriesColor;
+  color: Color;
   lineStyle: LineStyle;
 }
 
@@ -23,7 +23,7 @@ export function LinePreview({color, lineStyle}: Props) {
 
   const linePreviewColor = isGradientType(color)
     ? `url(#${gradientId.current})`
-    : getColorValue(color);
+    : color;
 
   return (
     <div>