+++ src/components/MultiSeriesBarChart/types.ts
@@ -5,10 +5,10 @@ import type {
   NumberLabelFormatter,
   DataSeries,
   Data,
-  SeriesColor,
+  Color,
 } from '../../types';
 
-export type Series = DataSeries<Data, SeriesColor>;
+export type Series = DataSeries<Data, Color>;
 
 export type StackSeries = ShapeSeries<
   {
@@ -19,7 +19,7 @@ export type StackSeries = ShapeSeries<
 
 export interface RenderTooltipContentData {
   data: {
-    color: SeriesColor;
+    color: Color;
     label: string;
     value: number;
   }[];
@@ -42,9 +42,6 @@ export enum BarMargin {
 }
 
 export interface BarOptions {
-  innerMargin: keyof typeof BarMargin;
-  outerMargin: keyof typeof BarMargin;
-  hasRoundedCorners: boolean;
   isStacked: boolean;
   /**
    * @deprecated This prop is experimental and not ready for general use. If you want to use this, come talk to us in #polaris-data-viz
@@ -52,23 +49,12 @@ export interface BarOptions {
   zeroAsMinHeight: boolean;
 }
 
-export interface GridOptions {
-  showHorizontalLines: boolean;
-  color: string;
-  horizontalOverflow: boolean;
-  horizontalMargin: number;
-}
-
 export interface XAxisOptions {
   labelFormatter: StringLabelFormatter;
-  showTicks: boolean;
   labels: string[];
-  labelColor: string;
 }
 
 export interface YAxisOptions {
   labelFormatter: NumberLabelFormatter;
-  labelColor: string;
-  backgroundColor: string;
   integersOnly: boolean;
 }