+++ src/components/SquareColorPreview/SquareColorPreview.tsx
@@ -1,22 +1,18 @@
 import React from 'react';
-import type {SeriesColor} from 'types';
+import type {Color} from 'types';
 
-import {
-  getColorValue,
-  isGradientType,
-  createCSSGradient,
-} from '../../utilities';
+import {isGradientType, createCSSGradient} from '../../utilities';
 
 import styles from './SquareColorPreview.scss';
 
 export interface SquareColorPreviewProps {
-  color: SeriesColor;
+  color: Color;
 }
 
 export function SquareColorPreview({color}: SquareColorPreviewProps) {
   const background = isGradientType(color)
-    ? createCSSGradient(color)
-    : getColorValue(color);
+    ? createCSSGradient(color, 305)
+    : color;
 
   return <div className={styles.ColorPreview} style={{background}} />;
 }