+++ src/components/SquareColorPreview/stories/SquareColorPreview.stories.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {Story, Meta} from '@storybook/react';
 
 import {SquareColorPreview, SquareColorPreviewProps} from '../../../components';
+import {colorTeal} from '../../../constants';
 
 export default {
   title: 'Subcomponents/SquareColorPreview',
@@ -9,7 +10,7 @@ export default {
   argTypes: {
     color: {
       description:
-        'The color to be displayed in the square. [SeriesColor type definition.](https://github.com/Shopify/polaris-viz/blob/master/src/types.ts#L2)',
+        'The CSS color or gradient array color to be displayed in the square.',
     },
   },
   parameters: {
@@ -35,7 +36,7 @@ const Template: Story<SquareColorPreviewProps> = (
 export const Solid = Template.bind({});
 
 Solid.args = {
-  color: 'primary',
+  color: colorTeal,
 };
 
 const purple = '#5052b3';