+++ src/components/StackedAreaChart/components/StackedAreas/tests/StackedAreas.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
 import {scaleLinear} from 'd3-scale';
-import type {Color} from 'types';
 
 import {StackedAreas} from '../StackedAreas';
 
@@ -18,8 +17,7 @@ describe('<StackedAreas />', () => {
     width: 100,
     height: 50,
     transform: '',
-    colors: ['colorRed', 'colorPurple'] as Color[],
-    opacity: 1,
+    colors: ['red', 'purple'],
     xScale: scaleLinear(),
     yScale: scaleLinear(),
     isAnimated: true,
@@ -92,10 +90,9 @@ describe('<StackedAreas />', () => {
     expect(stackedArea).toContainReactComponent('path', {
       // eslint-disable-next-line id-length
       d: 'M250,250L250,250L250,250L250,250Z',
-      fill: 'rgb(222, 54, 24)',
-      stroke: 'rgb(222, 54, 24)',
+      fill: 'url(#area-stackedAreas-1-0)',
+      stroke: 'url(#area-stackedAreas-1-0)',
       strokeWidth: '0.1',
-      opacity: 1,
     });
   });
 });